+++ Plugins~/Src/MeshSyncClientBlender/msblenGeometryNodeUtils.h
