+++ Plugins~/Src/MeshSyncClientBlender/msblenGeometryNodesUtils.cpp
@@ -1,4 +1,4 @@
-#include "msblenGeometryNodes.h"
+#include "msblenGeometryNodeUtils.h"
 #include <sstream>
 #include <BlenderPyObjects/BlenderPyContext.h>
 #include <BlenderPyObjects/BlenderPyDepsgraphObjectInstance.h>
